# api/notificationservice/notificationservice_swagger.yaml
basePath: /api/notification-service
definitions:
...
  maildto.CheckMailServerEntityRequest:
    properties:
      domain:
        type: string
      isAuthenticationRequired:
        default: false
        type: boolean
      isTlsEnforced:
        default: false
        type: boolean
      password:
        type: string
      port:
        type: integer
      username:
        type: string
    type: object
  maildto.CheckMailServerRequest:
    properties:
      domain:
        type: string
      isAuthenticationRequired:
        default: false
        type: boolean
      isTlsEnforced:
        default: false
        type: boolean
      password:
        type: string
      port:
        type: integer
      username:
        type: string
    type: object
  maildto.MailNotificationChannelRequest:
    properties:
      channelName:
        type: string
      domain:
        type: string
      id:
        type: string
      isAuthenticationRequired:
        default: false
        type: boolean
      isTlsEnforced:
        default: false
        type: boolean
      maxEmailAttachmentSizeMb:
        type: integer
      maxEmailIncludeSizeMb:
        type: integer
      password:
        type: string
      port:
        type: integer
      senderEmailAddress:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API of the Notification service
  license:
    name: AGPL-3.0-or-later
  title: Notification Service API
  version: "1.0"
paths:
...
  /notification-channel/mail:
    get:
      description: List mail notification channels by type
      parameters:
      - description: Channel type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/maildto.MailNotificationChannelRequest'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: List Mail Channels
      tags:
      - mail-channel
    post:
      consumes:
      - application/json
      description: Create a new mail notification channel
      parameters:
      - description: Mail channel to add
        in: body
        name: MailChannel
        required: true
        schema:
          $ref: '#/definitions/maildto.MailNotificationChannelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/maildto.MailNotificationChannelRequest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: Create Mail Channel
      tags:
      - mail-channel
  /notification-channel/mail/{id}:
    delete:
      description: Delete a mail notification channel
      parameters:
      - description: Mail channel ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Deleted successfully
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: Delete Mail Channel
      tags:
      - mail-channel
    put:
      consumes:
      - application/json
      description: Update an existing mail notification channel
      parameters:
      - description: Mail channel ID
        in: path
        name: id
        required: true
        type: string
      - description: Mail channel to update
        in: body
        name: MailChannel
        required: true
        schema:
          $ref: '#/definitions/maildto.MailNotificationChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/maildto.MailNotificationChannelRequest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: Update Mail Channel
      tags:
      - mail-channel
...
  /notifications/mail/{id}/check:
    post:
      consumes:
      - application/json
      description: Check if a mail server is reachable
      parameters:
      - description: Mail server to check
        in: body
        name: MailServerConfig
        required: true
        schema:
          $ref: '#/definitions/maildto.CheckMailServerEntityRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Mail server reachable
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Mail server error
      security:
      - KeycloakAuth: []
      summary: Check mail server
      tags:
      - mailserver
  /notifications/mail/check:
    post:
      consumes:
      - application/json
      description: Check if a mail server is reachable
      parameters:
      - description: Mail server to check
        in: body
        name: MailServerConfig
        required: true
        schema:
          $ref: '#/definitions/maildto.CheckMailServerRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Mail server reachable
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Mail server error
      security:
      - KeycloakAuth: []
      summary: Check mail server
      tags:
      - mailserver
...
securityDefinitions:
  KeycloakAuth:
    authorizationUrl: '{{.KeycloakAuthUrl}}/realms/{{.KeycloakRealm}}/protocol/openid-connect/auth'
    flow: implicit
    type: oauth2
swagger: "2.0"